chore(broker-core): move incident record to protocol
