+++ protocol-impl/src/main/java/io/zeebe/protocol/impl/record/value/incident/IncidentRecord.java
@@ -1,23 +1,20 @@
 /*
- * Zeebe Broker Core
  * Copyright © 2017 camunda services GmbH (info@camunda.com)
  *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
  *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
+ *     http://www.apache.org/licenses/LICENSE-2.0
  *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
-package io.zeebe.broker.incident.data;
+package io.zeebe.protocol.impl.record.value.incident;
 
-import io.zeebe.broker.logstreams.processor.TypedRecord;
 import io.zeebe.msgpack.UnpackedObject;
 import io.zeebe.msgpack.property.EnumProperty;
 import io.zeebe.msgpack.property.LongProperty;
@@ -115,13 +112,12 @@ public class IncidentRecord extends UnpackedObject {
   }
 
   public IncidentRecord initFromWorkflowInstanceFailure(
-      TypedRecord<WorkflowInstanceRecord> workflowInstanceEvent) {
-    final WorkflowInstanceRecord value = workflowInstanceEvent.getValue();
+      long key, WorkflowInstanceRecord workflowInstanceEvent) {
 
-    setElementInstanceKey(workflowInstanceEvent.getKey());
-    setBpmnProcessId(value.getBpmnProcessId());
-    setWorkflowInstanceKey(value.getWorkflowInstanceKey());
-    setElementId(value.getElementId());
+    setElementInstanceKey(key);
+    setBpmnProcessId(workflowInstanceEvent.getBpmnProcessId());
+    setWorkflowInstanceKey(workflowInstanceEvent.getWorkflowInstanceKey());
+    setElementId(workflowInstanceEvent.getElementId());
 
     return this;
   }
